SCRUM-125 feat: add useInitializeProfile hook for profile and posts initialization, refactor Profile component to utilize new hook
